Move json_path::v2 code into json_path
